requirements/requirements.txt
alembic
databases[postgresql]
graphene
sqlalchemy-utils
starlette
uvicorn
